challenges/pyproject.toml
[tool.poetry]
name = "clase9-pda"
version = "0.1.0"
description = ""
authors = ["Flor"]

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^1.3.3"
python-decouple = "^3.5"
fastapi = "^0.104.1"
python-multipart = "^0.0.6"
uvicorn = "^0.24.0"
SQLAlchemy = "^2.0.23"


[tool.poetry.dev-dependencies]
pdbpp = "^0.10.3"
pylint = "^2.11.1"
flake8 = "^4.0.1"
black = "^21.9b0"
flake8-docstrings = "^1.6.0"
mypy = "^0.910"
memory-profiler = "^0.58.0"
Sphinx = "^4.2.0"
pre-commit = "^3.5.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
